# Improve dependency validator's performance